Renamed ArraySerialization -> ContainerSerialization
